Static linking changed to dynamic
Classes are no longer exposed, instead a factory for the
JCON_NAMESPACE::console class has been added.
SFML is now dynamically linked so as to allow SFML to be used outside of
the jconsole.dll file, without creating linking errors due to multiple
MFC instances in the SFML library.
Premake5.lua has been altered to allow for this, aswell as simplifying
the file.